Implement boost::context as a fallback for non-blocking API support
The non-blocking API has native (assembler) implementations for x86_64,
i386, and (with recent patch) aarch64; these implementations are the most
efficient. For other architectures, a fallback to ucontext is supported.

But ucontext is not the most efficient, and it is not available on all
platforms (it has been deprecated in POSIX). The boost::context library
provides an alternative fallback that is available on more architectures and
should be more efficient than ucontext (if still not quite as fast as the
native support).

This patch adds boost::context as a dependency of libmariadb that provides a
fallback on non-natively supported architectures and is preferred over
ucontext, when both are available.

One problem with this patch is that it introduces a dependency on C++ and on
the boost library. This is rather intrusive on a library that is otherwise
pure C with few dependencies. It would be preferable to avoid this
dependency when it is not needed. Maybe boost::context should be an optional
feature enabled by the user when running cmake, and some cmake magic used to
not introduce the C++ dependency into the project unless explicitly
requested by the user.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>